move content, footer, header components from organisms folder to components
