+++ src/App.js
@@ -2,9 +2,9 @@ import React from 'react';
 
 import './App.scss';
 
-import Header from './Components/organisms/Header';
-import Content from './Components/organisms/Content';
-import Footer from './Components/organisms/Footer';
+import Header from './Components/Header';
+import Content from './Components/Content';
+import Footer from './Components/Footer';
 import Home from './Components/Pages/Home/';
 
 function App() {
