+++ src/components/Footer/Footer.test.js
