+++ src/components/Header/Header.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { Link } from '@reach/router';
 
 import Image from 'react-bootstrap/Image';
-import logo from '../../../images/logo-placeholder.svg';
+import logo from '../../images/logo-placeholder.svg';
 import './Header.scss';
 
 function Header() {
