+++ src/components/Header/Header.scss
@@ -1,4 +1,4 @@
-@import '../../../styles/variables';
+@import '../../styles/variables';
 
 .header {
   position: sticky;
